[#654] review: address review comments
